+++ src/components/common/form/ProfileImageButton.tsx
@@ -2,7 +2,7 @@
  * 프로필 사진 추가와 변경에 사용하는 공용 버튼 컴포넌트입니다.
  */
 
-import Button from '@/components/common/button/Button';
+import { Button } from '@/components/common/button';
 
 export default function ProfileImageButton() {
   // TODO: 프로필 이미지 업로드/변경 트리거 버튼 스타일과 접근성 라벨을 구현합니다.
