+++ src/components/common/toast/index.ts
@@ -1,2 +1,6 @@
 export { default as Toast } from '@/components/common/toast/components/Toast';
+export {
+  ToastProvider,
+  useToastContext,
+} from '@/components/common/toast/ToastProvider';
 export { useToast } from '@/components/common/toast/hooks/useToast';
